+++ forge/src/main/java/io/github/mortuusars/scholar/forge/event/CommonEvents.java
@@ -3,8 +3,11 @@ package io.github.mortuusars.scholar.forge.event;
 import io.github.mortuusars.scholar.Config;
 import io.github.mortuusars.scholar.Scholar;
 import io.github.mortuusars.scholar.network.forge.PacketsImpl;
+import io.github.mortuusars.scholar.world.entity.Reading;
 import net.minecraft.core.cauldron.CauldronInteraction;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.item.Items;
+import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -29,6 +32,13 @@ public class CommonEvents {
         }
     }
 
+    @Mod.EventBusSubscriber(modid = Scholar.ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
     public static class ForgeBus {
+        @SubscribeEvent
+        public static void onPlayerLogin(PlayerEvent.PlayerLoggedInEvent event) {
+            if (event.getEntity() instanceof ServerPlayer player) {
+                Reading.closeAllBooksInInventory(player);
+            }
+        }
     }
 }